Additional changes by Jeremy
I found a couple of problems with the changes that I made. This version
does not store the connection object. It connects with each query, but
it still only copies the database when necessary. I didn't mention this,
but the reason, why I made this change is that the plugin doesn't work
when firefox is running due to the fact that firefox locks the places
database.